requirements.txt
streamlit==1.44.1
google-.generativeai==0.8.4
dotenv==0.9.9
pillow==10.2.0